+++ gserver/src/steps.handler.ts
@@ -355,6 +355,12 @@ export default class StepsHandler {
                     res = res.replace(match[i], () => '${' + num + ':}');
                 }
             }
+
+            const favQuote = this.settings.cucumberautocomplete.favoriteQuoteChar;
+            if (favQuote){
+                res = res.replace("(\"|')", favQuote);
+                res = res.replace("\\1", favQuote);
+            }
         } else {
             //We can replace some outputs, ex. strings in brackets to make insert strings more neat
             res = res.replace(/\"\[\^\"\]\+\"/g, '""');