+++ gserver/src/typings.d.ts
@@ -28,6 +28,7 @@ interface Settings {
         onTypeFormat?: boolean,
         gherkinDefinitionPart?: string,
         stepRegExSymbol?: string
-        pureTextSteps?: boolean
+        pureTextSteps?: boolean, 
+        favoriteQuoteChar?: string
     }
 }