+++ gserver/test/data/steps/pureTextSteps.steps.js
@@ -1,1 +1,4 @@
 this.When('I give 3/4 and 5$')
+
+
+this.Then('I receive ^{string}$')